Split base-element.js into video-base-element.js
