+++ extensions/amp-video/1.0/base-element.js
@@ -13,60 +13,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import {ActionTrust} from '../../../src/core/constants/action-constants';
 import {CSS} from './autoplay.jss';
 import {PreactBaseElement} from '../../../src/preact/base-element';
 import {VideoWrapper} from './component';
 
-/** @extends {PreactBaseElement<VideoWrapperDef.Api>} */
-export class VideoBaseElement extends PreactBaseElement {
-  /** @override */
-  init() {
-    this.registerApiAction_('play', (api) => api.play());
-    this.registerApiAction_('pause', (api) => api.pause());
-    this.registerApiAction_('mute', (api) => api.mute());
-    this.registerApiAction_('unmute', (api) => api.unmute());
-
-    // Ugly that this action has three names, but:
-    // - requestFullscreen for consistency with Element.requestFullscreen
-    // - fullscreenenter / fullscreen are for backwards compatibility
-    const requestFullscreen = (api) => api.requestFullscreen();
-    this.registerApiAction_('requestFullscreen', requestFullscreen);
-    this.registerApiAction_('fullscreenenter', requestFullscreen);
-    this.registerApiAction_('fullscreen', requestFullscreen);
-  }
-
-  /**
-   * @param {string} alias
-   * @param {function(!VideoWrapperDef.Api, !../../../src/service/action-impl.ActionInvocation)} handler
-   * @param {!../../../src/core/constants/action-constants.ActionTrust=} minTrust
-   * @private
-   */
-  registerApiAction_(alias, handler, minTrust = ActionTrust.HIGH) {
-    this.registerApiAction(
-      alias,
-      (api, invocation) => {
-        if (invocation.trust >= ActionTrust.HIGH) {
-          // TODO(alanorozco): There may be a better solution that doesn't
-          // require this method which is not standard in HTMLMediaElement, like
-          // potentially toggling `autoplay` instead.
-          api.userInteracted();
-        }
-        handler(api, invocation);
-      },
-      minTrust
-    );
-  }
-}
+export class BaseElement extends PreactBaseElement {}
 
 /** @override */
-VideoBaseElement['Component'] = VideoWrapper;
+BaseElement['Component'] = VideoWrapper;
 
 /** @override */
-VideoBaseElement['loadable'] = true;
+BaseElement['loadable'] = true;
 
 /** @override */
-VideoBaseElement['layoutSizeDefined'] = true;
+BaseElement['layoutSizeDefined'] = true;
 
 /**
  * Defaults to `{component: 'video'}` from `VideoWrapper` component.
@@ -78,10 +38,10 @@ VideoBaseElement['layoutSizeDefined'] = true;
  * ```
  * @override
  */
-VideoBaseElement['staticProps'];
+BaseElement['staticProps'];
 
 /** @override */
-VideoBaseElement['props'] = {
+BaseElement['props'] = {
   'album': {attr: 'album'},
   'alt': {attr: 'alt'},
   'artist': {attr: 'artist'},
@@ -113,7 +73,7 @@ VideoBaseElement['props'] = {
 };
 
 /** @override */
-VideoBaseElement['shadowCss'] = CSS;
+BaseElement['shadowCss'] = CSS;
 
 /** @override */
-VideoBaseElement['usesShadowDom'] = true;
+BaseElement['usesShadowDom'] = true;
